Updated KAH
Fixed loading times for the bosses.
Re-balanced the gears and hp of both heros and bosses

PS: beware as a boss you should expect to get bullied if they are using the revolver. And remember that the boss fight scripts are just to have fun just laugh off being abused lol.